removed redundant classes and refactored
