+++ fetching-service/src/main/java/rs/edu/raf/banka1/repositories/InflationRepository.java
@@ -8,8 +8,4 @@ import java.util.List;
 
 @Repository
 public interface InflationRepository extends JpaRepository<Inflation, Long> {
-
-    List<Inflation> findAllByCurrencyId(Long currencyId);
-
-    List<Inflation> findAllByCurrencyIdAndYear(Long currencyId, Integer year);
 }
