+++ fetching-service/src/main/java/rs/edu/raf/banka1/services/ExchangeService.java
@@ -6,6 +6,4 @@ import java.util.List;
 
 public interface ExchangeService {
     void seedDatabase();
-    List<ExchangeDto> getAllExchanges();
-    ExchangeDto getExchangeById(Long id);
 }
