--- fetching-service/src/main/java/rs/edu/raf/banka1/services/MarketService.java
@@ -1,5 +1,0 @@
-package rs.edu.raf.banka1.services;
-
-public interface MarketService {
-    int exampleMethod();
-}
